requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.3.1
nltk>=3.6.3
flask>=2.0.1
transformers>=4.30.0
ultralytics>=8.0.0
python-dotenv>=0.19.0
pycocotools>=2.0.6
tqdm>=4.65.0
pandas>=1.3.3
matplotlib>=3.4.3
scikit-learn>=0.24.2 